requirements.txt
streamlit
pandas==2.1.0
numpy==1.22.4
plotly==5.6.0
wordcloud==1.9.3
transformers==4.25.1
torch==2.0.1
beautifulsoup4==4.11.1
nltk==3.9.1
scikit-learn
streamlit-option-menu
NRCLex==3.0.0


langchain-groq
groq
langchain_community
python-dotenv
langchain
langchain_core
langchain_huggingface
sqlalchemy
watchdog
pydantic
chromadb
